% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_alignment_loanbook_exposure.R
\name{aggregate_alignment_loanbook_exposure}
\alias{aggregate_alignment_loanbook_exposure}
\title{Return loan book level aggregation of company alignment metrics by exposure}
\usage{
aggregate_alignment_loanbook_exposure(
  data,
  matched,
  level = c("net", "bo_po"),
  .by = NULL
)
}
\arguments{
\item{data}{data.frame. Holds output pf company indicators}

\item{matched}{data.frame. Holds matched and prioritised loan book}

\item{level}{Character. Vector that indicates if the aggregate alignment
metric should be returned based on the net technology deviations (\code{net}) or
disaggregated into buildout and phaseout technologies (\code{bo_po}).}

\item{.by}{\if{html}{\out{<tidy-select>}} Character. Optionally, a selection of columns to
group by, instead. All columns indicated must be available variables in the
\code{matched} data set. The intended use case is to allow analyzing the loan
books by additional traits of interest, such as types of financial
institutions.}
}
\description{
Return loan book level aggregation of company alignment metrics by exposure
}
